Moved products.js to correct test data folder
